Revert "wallet: move experimental account graphs to future-features modal"
This reverts commit 3e4d844f103dec39a9d617a098860d9198efd11f.
