+++ app/tcoin/wallet/components/dashboard/ContributionsCard.tsx
@@ -9,7 +9,6 @@ export function ContributionsCard({
   charityData,
   openModal,
   closeModal,
-  compact = false,
 }: {
   selectedCharity: string;
   setSelectedCharity: (charity: string) => void;
@@ -20,7 +19,6 @@ export function ContributionsCard({
   };
   openModal: any;
   closeModal: any;
-  compact?: boolean;
 }) {
   return (
     <Card>
@@ -28,27 +26,23 @@ export function ContributionsCard({
         <CardTitle>Charitable Contributions</CardTitle>
       </CardHeader>
       <CardContent className="relative">
-        <div className={`relative h-full ${compact ? "space-y-1 text-sm" : "space-y-2"}`}>
+        <div className="space-y-2 relative h-full">
           <p>
             My default charity: <strong>{selectedCharity}</strong>
           </p>
           <p>
             My contribution to {selectedCharity}: {charityData.personalContribution} TCOIN
           </p>
-          {!compact && (
-            <>
-              <p>
-                All users to {selectedCharity}: {charityData.allUsersToCharity} TCOIN
-              </p>
-              <p>
-                All users to all charities: {charityData.allUsersToAllCharities} TCOIN
-              </p>
-            </>
-          )}
+          <p>
+            All users to {selectedCharity}: {charityData.allUsersToCharity} TCOIN
+          </p>
+          <p>
+            All users to all charities: {charityData.allUsersToAllCharities} TCOIN
+          </p>
         </div>
         <Button
           variant="default"
-          className={`bottom-0 left-0 h-auto py-1 font-normal ${compact ? "mt-2 text-xs" : ""}`}
+          className="py-1 h-auto bottom-0 left-0 font-normal"
           onClick={() => {
             openModal({
               content: (
