+++ app/tcoin/wallet/components/modals/index.ts
@@ -10,4 +10,3 @@ export * from "./BuyTcoinModal";
 export * from "./BiaPreferencesModal";
 export * from "./VoucherRoutingPreferencesModal";
 export * from "./ThemeSelectModal";
-export * from "./FutureAppFeaturesModal";
